manualstudio/backend/requirements.txt
# FastAPI and Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1

# Database
sqlalchemy>=2.0.23
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Celery and Redis
celery>=5.3.4
redis>=5.0.1

# S3/MinIO Storage
boto3>=1.34.0

# OpenAI
openai>=1.6.0

# Anthropic (optional)
anthropic>=0.8.0

# PPTX Generation
python-pptx>=0.6.23

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
jsonschema>=4.20.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.2
pillow>=10.1.0
